Update Exchange.sol
It seems the usdc value of the position/collateral is only calculated upon entry of the position, I would like it to be calculated anytime this smart contract requires to know the position value and the collateral value. This way if in the future a user ads non-ETH we only need to adjust the logic where it retrieves the non-ETH-USDC price from Uniswap. Furthermore, I would like to create first a variable for people on the liquidation list, with their liquidatable amounts in usdc. You need money to liquidate people, because you take over their position, so then we can use this list and start liquidating them with the discount. (also I dont see the liquidation discount here).